+++ trunk/GUIElements/ViewUnacknowledgedOrders.cs
@@ -132,7 +132,7 @@ namespace EveMarketMonitorApp.GUIElements
             // Only refresh the display if the number of unacknowledged orders has changed.
             OrdersList unack = Orders.LoadOrders(
                 UserAccount.CurrentGroup.GetAssetAccessParams(APIDataType.Orders),
-                new List<int>(), new List<int>(), (int)OrderState.ExpiredOrFilledAndUnacknowledged, "Any");
+                new List<int>(), new List<long>(), (int)OrderState.ExpiredOrFilledAndUnacknowledged, "Any");
             if (unack.Count != _lastNumberOfOrders)
             {
